Moving dialog windows to own module.
